Submodule of lessphp is referenced
This is to help the users to automatically checkout the code as is
without trying to figure out what other dependencies are required
to be added to the project path.

The .gitignore entry was removed as this is where the third party
dependency was supposed to sit. Now we can include it as we have
the symbolic link as a submodule and won't be committing other
people's code.